[#50] UIControl 방식 삭제, UIButton 방식으로 결정
