+++ YDS-Sample/PageListViewController.swift
@@ -38,7 +38,6 @@ class PageListViewController: UIViewController {
     
     let atomPages: [Page] = [
         Page(title: "Checkbox", vc: CheckboxPageViewController.self),
-        Page(title: "CheckboxBtnVer", vc: CheckboxBtnVerPageViewController.self),
         Page(title: "SimpleTextField", vc: SimpleTextFieldSampleViewController.self),
         Page(title: "SuffixTextField", vc: SuffixTextFieldSampleViewController.self),
         Page(title: "PasswordTextField", vc: PasswordTextFieldSampleViewController.self),
